refactor: restructure Token component into separate files and introduce TokenGroup component for better organization and usage
